aggiornamento delle classi del package campaign

+++ Backend/src/main/java/it/unicam/cs/ids/loyaltyPlatform/model/campaign/CampaignServiceImpl.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
-@Table()
+@Table(name = "")
 public class CampaignServiceImpl implements CampaignService {
 
     @Autowired
@@ -20,8 +20,8 @@ public class CampaignServiceImpl implements CampaignService {
 
 
     @Override
-    public void save(Campaign campaign) {
-        this.repository.save(campaign);
+    public Campaign save(Campaign campaign) {
+        return this.repository.save(campaign);
     }
 
     @Override
@@ -30,22 +30,41 @@ public class CampaignServiceImpl implements CampaignService {
         // repository doesn't provide a ready to use saveById method.
     }
 
+    @Override
     public Campaign update(@NonNull Campaign campaign) {
-        Campaign oldCampaign = repository.findById(campaign.getId()).orElse(null);
-        if (existById(oldCampaign.getId())) {
+        Campaign oldCampaign = this.repository.findById(campaign.getId()).orElse(null);
+
+        if (existsById(oldCampaign.getId())) {
             if (oldCampaign.getId().equals(campaign.getId())) {
                 return this.repository.save(campaign);
             } else throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "...");
         } else throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Inexistent campaign.");
     }
 
+    @Override
+    public Campaign updateById(@NonNull UUID id) {
+        Optional<Campaign> old = this.repository.findById(id);
+
+        if (old.isPresent()) {
+            Campaign originalCampaign = old.get();
+            return this.repository.save(originalCampaign);
+        }
+        return null;
+    }
+
+    @Override
     public Campaign findById(@NonNull UUID id) {
         return this.repository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Campaign not found"));
     }
 
     @Override
     public Optional<Campaign> getById(@NonNull UUID id) {
-        return this.repository.findById(id);
+        Optional<Campaign> campaign = this.repository.findById(id);
+        if (campaign.isPresent()) {
+            return this.repository.findById(id);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Campaign not found");
+        }
     }
 
     @Override
@@ -60,10 +79,15 @@ public class CampaignServiceImpl implements CampaignService {
 
     @Override
     public void deleteById(UUID id) {
-        this.repository.deleteById(id);
+        if (this.repository.findById(id).isPresent()) {
+            this.repository.deleteById(id);
+            System.out.println("Campaign deleted successfully");
+        } else {
+            System.out.println("No such campaign in the database");
+        }
     }
 
-    private boolean existById(@NonNull UUID id) {
+    private boolean existsById(@NonNull UUID id) {
         return this.repository.existsById(id);
     }
